Created PlacingWireMod system.
