+++ src/signalNetwork/BlockConnection.cs
@@ -84,10 +84,10 @@ namespace signals.src.transmission
         public override bool OnBlockInteractStart(IWorldAccessor world, IPlayer byPlayer, BlockSelection blockSel)
         {
 
-            HangingWiresMod mod = api.ModLoader.GetModSystem<HangingWiresMod>();
+            PlacingWiresMod mod = api.ModLoader.GetModSystem<PlacingWiresMod>();
             if (mod == null)
             {
-                api.Logger.Error("HangingWiresMod mod system not found");
+                api.Logger.Error("PlacingWiresMod mod system not found");
             }
             else
             {
